internship_bot/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-telegram-bot==20.8
sqlalchemy==2.0.25
python-dateutil==2.8.2
